handle enum correctly - cast to int

+++ src/Day-03-IntroToConditionalStatements/03-Solution.cs
@@ -91,7 +91,7 @@ class Result
             resultWeirdness = weirdness.EndWeirdness;
         }
 
-        Console.WriteLine(Result.getWeirdness(resultWeirdness));
+        Console.WriteLine(Result.getWeirdness((int)resultWeirdness));
     }
 }
